:construction: Reprojection and alignment script for mask and Sentinel-2
Reprojecting the high resolution (<2m) image mask and ensuring the pixels align with the Sentinel-2 (10m) reference. Relies on rioxarray for the reprojection and clipping, and xarray for the pixel alignment (via interpolation). Lots of hardcoding, with the high resolution mask being resampled to 2m (5x that of 10m). The hope is that all this messy code will eventually be replaced using torchgeo when it supports inputs with different projections and resolutions properly.