:hammer: Reproject and align highres RGB-NIR imagery too and handle NaNs
Pre-processing the highres 4-band optical imagery in addition to the 1-band mask. Lots of extra dtype casting to deal with NaN values in the 1-band mask properly. Specifically, doing the reprojection of the high resolution (2m) images in float32 instead of uint8/16. The 1-band mask is now also masked properly with NaNs for areas without any valid imagery, normalized to 0-1 range, and stored as a float32 GeoTiff instead of uint8 before. The 4-band imagery is stored as a uint16 GeoTiff. Also fixed some sorting bugs, and hope that this piece of code will run smoothly overnight!