:chart_with_upwards_trend: Track mean IoU and Dice coefficient metrics
Calculating some evaluation metrics to better monitor the performance of the segmentation results. Specifically, the Intersection over Union (Jaccard Index), and Dice coefficient (F1 Score). May decide to use the more general F-score once we decide on whether precision or recall is more important. Oh, and this uses mmsegmentation's built-in eval_metrics function which requires casting to numpy. Might be better to move to torchmetrics at some point, but they currently require the target mask to be in int for some reason.