:alien: Use torchmetrics instead of mmseg to calculate IoU and F1 score
Using a more robust metrics package! The mmseg.core.eval_metrics was giving really small IoU/Jaccard and Dice/F1 values (maybe because I'm not doing the averaging correctly) and only worked on the CPU. Moving to torchmetrics allows us to run things on the GPU and gives proper looking fraction values, at the expense of having to binarize the target mask (so that values <0.5 become 0 and >0.5 become 1). Currently calculating these two metrics separately, but may decide to use torchmetrics.MetricCollection in the future if we want more scores to judge the model's performance.