:stethoscope: Implement validation step to check for overfitting
Compute the validation loss and metrics on the validation data (20% of the full dataset). The loss and metric calculation is about the same, just the logging is different, so there is a shared evaluate class function that is now called in both the training and validation steps. Needed to switch to tensorboard's add_scalars (instead of add_scalar) so that both the training and validation metrics show up in the same plot, see https://github.com/PyTorchLightning/pytorch-lightning/issues/2563#issuecomment-656506419.