:ambulance: Fix ValueError due to some weird xarray coordinate mismatch
Specifically, fixing `ValueError: zero-size array to reduction operation minimum which has no identity`. Doing so by re-assigning the x and y coordinates of the binary label mask to be the same as the NaN mask. See https://gis.stackexchange.com/questions/409527/rioxarray-zero-size-array-to-reduction-operation-minimum-which-has-no-identity/409528#409528 and https://github.com/corteva/rioxarray/issues/298#issuecomment-820559379.