:globe_with_meridians: Create index tiles for each aligned GeoTIFF file
Produce a GeoJSON that stores the bounding box Polygon of each training image tile! Besides the filenames of the Sentinel-2, mask and high-resolution images, the datetime of when the Sentinel-2 image was acquired is also saved. Also included an if-then code-block I use to resume the data reprojection/alignment when things crash, and gitignoring GeoJSON files for now.