:monocle_face: Implement test_step on independent hold-out set of images
To allow fair comparison of different models, compute F1 and IoU metrics on 8 independent images from a hold-out set (see 00a1c5b4fa9992749221d3f02bbd8a1e99447890) that was handpicked to be from different geographic localities compared to the train/val set. The evaluate function now has a boolean `calc_loss` parameter, so that the expensive loss computation can be skipped during the `test_step` loop (i.e. compute metrics only).

Note on issues and workarounds. Computing the test metrics means the full-size masks need to be loaded, and for some reason the geographical extent of the mask doesn't align with the image so need to do `rio.clip_box`. Also, the predicted tensor shapes are sometimes not exactly 5x that of the input image, so need to interpolate to the right size (which isn't actually proper, I know). Also needed to use CPU for test_step because GPU did not have enough memory, and use 32-bit precision too.