:heavy_plus_sign: Add deepspeed
DeepSpeed is a deep learning optimization library that makes distributed training and inference easy, efficient, and effective! Using PyPI source for now until conda-forge package is released.

Also need to install newer gcc version to prevent error `Your compiler (c++ 4.8.5) may be ABI-incompatible with PyTorch! Please use a compiler that is ABI-compatible with GCC 5.0 and above` on the hpc server.