:children_crossing: Allow passing specific image ids to run inference on
Finer control on which Sentinel-2 images we want to predict on! Create an image_set parameter flag in S2S2Dataset that accepts either 'trainval', 'test' or 'predict'. Using 'trainval' loads the data from the npy files; 'test' loads the image and mask from the GeoTIFFs; and 'predict' just loads the image from the GeoTIFF. This way, the predict_dataloader will return non-cropped Sentinel-2 images. Also fixed a bug in the predict_step that didn't convert the dtype to float16 properly.