:wrench: Allow setting root directory in S2S2DataModule
Enable running inference on any folder with Sentinel-2 6band GeoTIFFs. The ids parameter can now be a list of GeoTIFF filepaths instead of just a list of folder names. Makes it easier for example to produce predictions on an entire time-series of Sentinel-2 GeoTIFF data stored in a single flat folder.